# scaffold/acme_root/clients/typescript/openapi.json
{
  "components": {
    "schemas": {
...
      "CrossingBody": {
        "additionalProperties": false,
        "description": "What the sender declares of the bytes it sends: what they are, their\nSHA-256 in hex, and their size.",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/CrossingKind"
          },
          "sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Sha256",
            "type": "string"
          },
          "size": {
            "minimum": 0.0,
            "title": "Size",
            "type": "integer"
          }
        },
        "required": [
          "kind",
          "sha256",
          "size"
        ],
        "title": "CrossingBody",
        "type": "object"
      },
      "CrossingKind": {
        "enum": [
          "enrollment",
          "claim",
          "stream_part",
          "artifact",
          "result"
        ],
        "title": "CrossingKind",
        "type": "string"
      },
...
      "ExecDetailView": {
        "description": "An item the host holds: the call, the workspace, and the operation.",
        "properties": {
          "call_id": {
            "format": "uuid",
            "title": "Call Id",
            "type": "string"
          },
          "deadline": {
            "format": "date-time",
            "title": "Deadline",
            "type": "string"
          },
          "effect": {
            "title": "Effect",
            "type": "string"
          },
          "epoch": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Epoch"
          },
          "item_id": {
            "format": "uuid",
            "title": "Item Id",
            "type": "string"
          },
          "location": {
            "title": "Location",
            "type": "string"
          },
          "org_id": {
            "format": "uuid",
            "title": "Org Id",
            "type": "string"
          },
          "request": {
            "additionalProperties": true,
            "title": "Request",
            "type": "object"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "spec": {
            "additionalProperties": true,
            "title": "Spec",
            "type": "object"
          }
        },
        "required": [
          "item_id",
          "call_id",
          "org_id",
          "session_id",
          "effect",
          "deadline",
          "epoch",
          "spec",
          "location",
          "request"
        ],
        "title": "ExecDetailView",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LeaseView": {
        "properties": {
          "lease_expires_at": {
            "format": "date-time",
            "title": "Lease Expires At",
            "type": "string"
          }
        },
        "required": [
          "lease_expires_at"
        ],
        "title": "LeaseView",
        "type": "object"
      },
...
      "OutputStream": {
        "enum": [
          "stdout",
          "stderr"
        ],
        "title": "OutputStream",
        "type": "string"
      },
...
      "PartRequest": {
        "additionalProperties": false,
        "description": "One part of an item's output, in the order the host read it.",
        "properties": {
          "crossing": {
            "$ref": "#/components/schemas/CrossingBody"
          },
          "data": {
            "maxLength": 90000,
            "title": "Data",
            "type": "string"
          },
          "seq": {
            "minimum": 0.0,
            "title": "Seq",
            "type": "integer"
          },
          "stream": {
            "$ref": "#/components/schemas/OutputStream"
          }
        },
        "required": [
          "seq",
          "stream",
          "data",
          "crossing"
        ],
        "title": "PartRequest",
        "type": "object"
      },
...
      "ResultRequest": {
        "additionalProperties": false,
        "description": "How an item ended: the JSON of an exec result, in base64.",
        "properties": {
          "crossing": {
            "$ref": "#/components/schemas/CrossingBody"
          },
          "data": {
            "maxLength": 16000000,
            "title": "Data",
            "type": "string"
          }
        },
        "required": [
          "data",
          "crossing"
        ],
        "title": "ResultRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Acme API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/hosts/me/exec/{item_id}": {
      "get": {
        "description": "What the host runs for an item it holds: the command or the file\noperation, opened, with its deadline, its writer epoch, and where the\nworkspace is.",
        "operationId": "detail_v1_hosts_me_exec__item_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-app",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App"
            }
          },
          {
            "in": "header",
            "name": "x-app-version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Version"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecDetailView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Detail",
        "tags": [
          "relay"
        ]
      }
    },
    "/v1/hosts/me/exec/{item_id}/lease": {
      "post": {
        "description": "Renews the host's lease on the item while it runs.",
        "operationId": "extend_v1_hosts_me_exec__item_id__lease_post",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-app",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App"
            }
          },
          {
            "in": "header",
            "name": "x-app-version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Version"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaseView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Extend",
        "tags": [
          "relay"
        ]
      }
    },
    "/v1/hosts/me/exec/{item_id}/parts": {
      "post": {
        "description": "One part of the item's output. A part sent again lands once.",
        "operationId": "push_part_v1_hosts_me_exec__item_id__parts_post",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-app",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App"
            }
          },
          {
            "in": "header",
            "name": "x-app-version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Version"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Push Part",
        "tags": [
          "relay"
        ]
      }
    },
    "/v1/hosts/me/exec/{item_id}/result": {
      "post": {
        "description": "How the item ended, stored under its call's key. The first\nsettlement wins: one its lease or a stop settled already is refused.",
        "operationId": "push_result_v1_hosts_me_exec__item_id__result_post",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-app",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App"
            }
          },
          {
            "in": "header",
            "name": "x-app-version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Version"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResultRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Push Result",
        "tags": [
          "relay"
        ]
      }
    },
...
  }
}